+++ src/Board.h
@@ -9,8 +9,8 @@ public:
 
     ~Board();
 
-    const std::vector<Tile> &operator[](int i) const;
-    std::vector<Tile> &operator[](int i);
+    const std::vector<Tile> &operator[](size_t i) const;
+    std::vector<Tile> &operator[](size_t i);
 
 private:
     std::vector<std::vector<Tile>> board;
